feat(analyzers): Add analyzer for SerializeCollection TypeIdType mismatch
Adds a new Roslyn analyzer to validate polymorphic collection serialization.

This analyzer, `SerializeCollectionSingleTypeIdAnalyzer`, ensures that when `[SerializeCollection]` is used with `PolymorphicMode.SingleTypeId`, the `TypeIdType` property matches the type of the ID in the `[PolymorphicOption]` attributes.

- If `TypeIdType` is explicitly set, its type must match the type of the ID in `[PolymorphicOption]`.
- If `TypeIdType` is not set, no check is performed.

A diagnostic error (FSG1014) is reported if there is a mismatch.

Includes:
- The new analyzer implementation.
- Unit tests to verify correct and incorrect usage.
- Resource strings for the new diagnostic message.
- A fix in an existing use case to comply with the new rule.

+++ src/FourSer.Consumer/UseCases/SingleTypeIdTest.cs
@@ -8,8 +8,8 @@ namespace FourSer.Consumer.UseCases
         public byte AnimalType { get; set; }
 
         [SerializeCollection(PolymorphicMode = PolymorphicMode.SingleTypeId, TypeIdType = typeof(byte), TypeIdProperty = nameof(AnimalType))]
-        [PolymorphicOption(1, typeof(CatBase))]
-        [PolymorphicOption(2, typeof(DogBase))]
+        [PolymorphicOption((byte)1, typeof(CatBase))]
+        [PolymorphicOption((byte)2, typeof(DogBase))]
         public List<AnimalBase> Animals { get; set; } = new();
     }
